feat: 알림 기능 구현 (#365)
* feat: Notification 이벤트 리스너 구현

- Entity, Repository 구현
- 이벤트 객체 구현
- 이벤트 핸들러 구현

* feat: Async와 TransactionEventListener를 통해 NotificationEventHandler 구현

- 하나의 쓰레드는 하나의 트랜잭션만을 가지지 않을까하여 작성

* test: 최신 기술 조회 테스트 코드 작성

* feat: 댓글 등록/수정 시 알림 기능 구현

* feat: 알림 조회 기능 구현

* feat: 알림 단일 삭제/일괄 삭제 구현

* refactor: 알림 조회 시 최신순 정렬

* refactor: 로그인 유저의 알림 필드 추가

* feat: notification 테이블 추가한 sql문 추가

+++ backend/src/main/java/com/wooteco/nolto/user/ui/dto/ProfileResponse.java
@@ -15,10 +15,10 @@ public class ProfileResponse {
     private String nickname;
     private String bio;
     private String imageUrl;
-    private int notifications;
+    private long notifications;
     private LocalDateTime createdAt;
 
-    public static ProfileResponse of(User user, int notifications) {
+    public static ProfileResponse of(User user, long notifications) {
         return new ProfileResponse(user.getId(), user.getNickName(), user.getBio(), user.getImageUrl(),
                 notifications, user.getCreatedDate());
     }
